+++ common/retry_test.go
@@ -1,4 +1,4 @@
-package workspace
+package common
 
 import (
 	"context"
@@ -11,33 +11,39 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
-func TestRobustGetStatus_NoError(t *testing.T) {
+func TestRetryOnTimeout_NoError(t *testing.T) {
 	w := mocks.NewMockWorkspaceClient(t)
 	expected := &workspace.ObjectInfo{}
 	api := w.GetMockWorkspaceAPI().EXPECT()
 	api.GetStatusByPath(mock.Anything, mock.Anything).Return(expected, nil)
-	res, err := robustGetStatus(context.Background(), w.WorkspaceClient, "path")
+	res, err := RetryOnTimeout(context.Background(), func(ctx context.Context) (*workspace.ObjectInfo, error) {
+		return w.WorkspaceClient.Workspace.GetStatusByPath(ctx, "path")
+	})
 	assert.NoError(t, err)
 	assert.Equal(t, expected, res)
 }
 
-func TestRobustGetStatus_OneError(t *testing.T) {
+func TestRetryOnTimeout_OneError(t *testing.T) {
 	w := mocks.NewMockWorkspaceClient(t)
 	expected := &workspace.ObjectInfo{}
 	api := w.GetMockWorkspaceAPI().EXPECT()
 	call1 := api.GetStatusByPath(mock.Anything, mock.Anything).Return(nil, errors.New("request failed: request timed out after 1m0s of inactivity"))
 	call1.Repeatability = 1
 	api.GetStatusByPath(mock.Anything, mock.Anything).Return(expected, nil)
-	res, err := robustGetStatus(context.Background(), w.WorkspaceClient, "path")
+	res, err := RetryOnTimeout(context.Background(), func(ctx context.Context) (*workspace.ObjectInfo, error) {
+		return w.WorkspaceClient.Workspace.GetStatusByPath(ctx, "path")
+	})
 	assert.NoError(t, err)
 	assert.Equal(t, expected, res)
 }
 
-func TestRobustGetStatus_NonRetriableError(t *testing.T) {
+func TestRetryOnTimeout_NonRetriableError(t *testing.T) {
 	w := mocks.NewMockWorkspaceClient(t)
 	expected := errors.New("request failed: non-retriable error")
 	api := w.GetMockWorkspaceAPI().EXPECT()
 	api.GetStatusByPath(mock.Anything, mock.Anything).Return(nil, expected)
-	_, err := robustGetStatus(context.Background(), w.WorkspaceClient, "path")
+	_, err := RetryOnTimeout(context.Background(), func(ctx context.Context) (*workspace.ObjectInfo, error) {
+		return w.WorkspaceClient.Workspace.GetStatusByPath(ctx, "path")
+	})
 	assert.ErrorIs(t, err, expected)
 }
