+++ workspace/data_directory.go
@@ -18,7 +18,9 @@ func DataSourceDirectory() common.Resource {
 		WorkspacePath string `json:"workspace_path,omitempty" tf:"computed"`
 	}
 	return common.WorkspaceData(func(ctx context.Context, d *Directory, client *databricks.WorkspaceClient) error {
-		data, err := robustGetStatus(ctx, client, d.Path)
+		data, err := common.RetryOnTimeout(ctx, func(ctx context.Context) (*workspace.ObjectInfo, error) {
+			return client.Workspace.GetStatusByPath(ctx, d.Path)
+		})
 		if err != nil {
 			return err
 		}
